Added precedence to the parser and a repr to BinOp
The __repr__ method in Binary Operation is a quality of life thing that
will make debugging a hell of a lot easier. Precedence in parser.py will
allow me to implement order of operations but the code will be much
easier to follow/read and maintain. I also changed the expression rule
to be able to select the correct Operator